Implemented unit tests for current data.
The basis for allowing unit tests is to be able to work on non-file data. Therefore `IStreamProvider` interface is introduced that can represent memory streams during unit tests.

Note: to simplify the transition to RRB sprites (where commands and options will change), this commit doesn't allow setting up output folder and naming template. We'll prepare individual options for each export later on, but for the moment `--out-folder` and `--out-name` were removed.