Implemented RRB sprites handling.
With this the whole command handling was refactored - for RRB sprites we need to run all commands in several steps since this data is appended to the end of screens.

Also moved VisualStudio project and example projects into separate sub-directories. And introducing Acme assembler example - it's supported out of the box by m65dbg with source debugging which is just so much nicer than having to alt-tab between source and debugger all the time...